requirements.txt
colorama==0.4.3
pythonwhois-alt==2.4.4
RashlyOutlaid==0.11.1
tabulate==0.8.7
